backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
chromadb
python-docx
pymupdf
pydantic
python-dotenv
python-multipart
